# dist/swagger-output.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "REST API",
        "description": ""
    },
    "host": "localhost:3000",
    "basePath": "/",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/api/v1/farmer/": {
            "get": {
                "description": "",
                "parameters": [
                    {
                        "name": "authorization",
                        "in": "header",
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/api/v1/prosperity-hub/": {
            "post": {
                "description": "",
                "parameters": [
                    {
                        "name": "authorization",
                        "in": "header",
                        "type": "string"
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Bad Request"
                    }
                }
            },
            "get": {
                "description": "",
                "parameters": [
                    {
                        "name": "authorization",
                        "in": "header",
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/api/v1/warehouse/": {
            "post": {
                "description": "",
                "parameters": [
                    {
                        "name": "authorization",
                        "in": "header",
                        "type": "string"
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Bad Request"
                    }
                }
            },
            "get": {
                "description": "",
                "parameters": [
                    {
                        "name": "authorization",
                        "in": "header",
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/api/v1/farm/": {
            "post": {
                "description": "",
                "parameters": [
                    {
                        "name": "authorization",
                        "in": "header",
                        "type": "string"
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Bad Request"
                    }
                }
            },
            "get": {
                "description": "",
                "parameters": [
                    {
                        "name": "authorization",
                        "in": "header",
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/api/v1/wallet/": {
            "post": {
                "description": "",
                "parameters": [
                    {
                        "name": "authorization",
                        "in": "header",
                        "type": "string"
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Bad Request"
                    }
                }
            },
            "get": {
                "description": "",
                "parameters": [
                    {
                        "name": "authorization",
                        "in": "header",
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    }
}